stage-6: move exceptions to separate package
