+++ src/main/java/ru/vk/itmo/viktorkorotkikh/exceptions/BackgroundExecutionException.java
@@ -1,4 +1,4 @@
-package ru.vk.itmo.viktorkorotkikh;
+package ru.vk.itmo.viktorkorotkikh.exceptions;
 
 public class BackgroundExecutionException extends RuntimeException {
     public BackgroundExecutionException(Throwable cause) {
