+++ src/main/java/ru/vk/itmo/viktorkorotkikh/exceptions/CompactionException.java
@@ -1,4 +1,4 @@
-package ru.vk.itmo.viktorkorotkikh;
+package ru.vk.itmo.viktorkorotkikh.exceptions;
 
 public class CompactionException extends RuntimeException {
     public CompactionException(Throwable cause) {
