+++ src/main/java/ru/vk/itmo/viktorkorotkikh/exceptions/FlushingException.java
@@ -1,4 +1,4 @@
-package ru.vk.itmo.viktorkorotkikh;
+package ru.vk.itmo.viktorkorotkikh.exceptions;
 
 public class FlushingException extends RuntimeException {
     public FlushingException(Throwable cause) {
