+++ src/main/java/ru/vk/itmo/viktorkorotkikh/exceptions/LSMDaoCreationException.java
@@ -1,4 +1,4 @@
-package ru.vk.itmo.viktorkorotkikh;
+package ru.vk.itmo.viktorkorotkikh.exceptions;
 
 public class LSMDaoCreationException extends RuntimeException {
     public LSMDaoCreationException(Throwable cause) {
