+++ src/main/java/ru/vk/itmo/viktorkorotkikh/exceptions/LSMDaoOutOfMemoryException.java
@@ -1,4 +1,4 @@
-package ru.vk.itmo.viktorkorotkikh;
+package ru.vk.itmo.viktorkorotkikh.exceptions;
 
 public class LSMDaoOutOfMemoryException extends RuntimeException {
     public LSMDaoOutOfMemoryException() {
