+++ src/main/java/ru/vk/itmo/viktorkorotkikh/exceptions/TooManyFlushesException.java
@@ -1,4 +1,4 @@
-package ru.vk.itmo.viktorkorotkikh;
+package ru.vk.itmo.viktorkorotkikh.exceptions;
 
 public class TooManyFlushesException extends RuntimeException {
 }
